Splitted maps functions and options into several files.
