+++ ts/Maps/MapNavigationOptions.d.ts
@@ -1,0 +1,68 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type {
+    AlignValue,
+    VerticalAlignValue
+} from '../Core/Renderer/AlignObject';
+import type CSSObject from '../Core/Renderer/CSSObject';
+import type SVGAttributes from '../Core/Renderer/SVG/SVGAttributes';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export type ButtonRelativeToValue = ('plotBox'|'spacingBox');
+
+export interface MapNavigationButtonOptions {
+    align?: AlignValue;
+    alignTo?: ButtonRelativeToValue;
+    height?: number;
+    onclick?: Function;
+    padding?: number;
+    style?: CSSObject;
+    text?: string;
+    theme?: SVGAttributes;
+    verticalAlign?: VerticalAlignValue;
+    width?: number;
+    x?: number;
+    y?: number;
+}
+
+export interface MapNavigationOptions {
+    buttonOptions?: MapNavigationButtonOptions;
+    buttons?: Record<string, MapNavigationButtonOptions>;
+    enableButtons?: boolean;
+    enabled?: boolean;
+    enableDoubleClickZoom?: boolean;
+    enableDoubleClickZoomTo?: boolean;
+    enableMouseWheelZoom?: boolean;
+    enableTouchZoom?: boolean;
+    mouseWheelSensitivity?: number;
+}
+
+// @todo declare module
+declare global {
+    namespace Highcharts {
+        interface Options {
+            mapNavigation?: MapNavigationOptions;
+        }
+    }
+}
+
+export default MapNavigationOptions;
